refactor(rustc_lint): inline `check_lint_name_cmdline`
